agent_lab/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-lab"
version = "0.1.0"
description = "Local MLOps platform for evaluating LangGraph agents"
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.12",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "langfuse>=2.36",
    "langchain-openai>=0.1",
    "langgraph>=0.1",
    "langchain>=0.2",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "openai>=1.30",
    "langchain-benchmarks==0.0.14",
    "ragas>=0.2",
    "faiss-cpu>=1.7",
    "langchain-community>=0.2",
]

[project.scripts]
agentlab = "agent_lab_core.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_lab_core*"]
